server/requirements.txt
# Core LLM framework
llama-index-core>=0.10.0,<0.12.0
llama-index-utils-workflow
llama-index-readers-file

# Vector store
chromadb
llama-index-vector-stores-chroma

# Ollama integrations
llama-index-embeddings-ollama
llama-index-llms-ollama

# Code parsing
tree-sitter>=0.21.0,<0.22.0
tree-sitter-languages>=1.10.0

# API framework
fastapi
uvicorn[standard]
python-dotenv
